joineRML: Joint Modelling of Multivariate Longitudinal Data and Time-to-Event Outcomes

Fits the joint model proposed by Henderson and colleagues (2000) doi:10.1093/biostatistics/1.4.465 , but extended to the case of multiple continuous longitudinal measures. The time-to-event data is modelled using a Cox proportional hazards regression model with time-varying covariates. The multiple longitudinal outcomes are modelled using a multivariate version of the Laird and Ware linear mixed model. The association is captured by a multivariate latent Gaussian process. The model is estimated using a Monte Carlo Expectation Maximization algorithm. This project was funded by the Medical Research Council (Grant number MR/M013227/1).

See also

Author

Maintainer: Graeme L. Hickey graemeleehickey@gmail.com (ORCID)

Authors:

Pete Philipson peter.philipson1@newcastle.ac.uk (ORCID)

Ruwanthi Kolamunnage-Dona kdrr@liverpool.ac.uk (ORCID)

Alessandro Gasparini alessandro.gasparini@ki.se (ORCID)

Other contributors:

Andrea Jorgensen aljorgen@liverpool.ac.uk (ORCID) [contributor]

Paula Williamson p.r.williamson@liverpool.ac.uk (ORCID) [contributor]

Dimitris Rizopoulos d.rizopoulos@erasmusmc.nl (data/renal.rda, R/hessian.R, R/vcov.R) [contributor, data contributor]

Medical Research Council (Grant number: MR/M013227/1) [funder]